Another big step.
We are now creating the equivalent of emails (with \n) that just need to be turned into actual emails.

In order to empower this, we need a couple more steps where we add the email to the DB, and then set up an email in php powered by that information.

Also, turned a lot of the copy/paste into functions.